[joint commit optimizer-ortools] Moving quantity division from optimizer-ortools to optimizer-api because multiplication is in optimizer-api but we can move this float->int transformation to optimizer-ortools because rest of the trasnformation is done there (i.e., all cost trasnformations).
Fixing result display of the cost

Fix filters_test precision_coef test accordingly

+++ wrappers/ortools_result_pb.rb
@@ -16,7 +16,7 @@ Google::Protobuf::DescriptorPool.generated_pool.build do
     repeated :activities, :message, 1, "ortools_result.Activity"
   end
   add_message "ortools_result.Result" do
-    optional :cost, :int64, 1
+    optional :cost, :float, 1
     optional :duration, :float, 2
     optional :iterations, :int32, 3
     repeated :routes, :message, 4, "ortools_result.Route"
